+++ CHANGELOG.md
@@ -3,9 +3,15 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 
-## [0.1.1](https://github.com/wilspi/django-template-editor/releases/tag/v0.1.1) - 2020-02-08
+## [0.1.2](https://github.com/wilspi/django-templatestore/releases/tag/v0.1.2) - 2020-03-05
 ### Added
--  Reusable django application: `django-template-editor`.
+-  Schema to store templates and resp. versions.
+-  APIs to support storage and editing templates, read complete documentation [here](https://raw.githubusercontent.com/wilspi/django-templatestore/0.1.2/docs/swagger.json).
+
+
+## [0.1.1](https://github.com/wilspi/django-templatestore/releases/tag/v0.1.1) - 2020-02-08
+### Added
+-  Reusable django application: `django-templatestore`.
 -  Pure API that returns rendered template given template and context data.
 -  UI: Template and json web editors, and text view for rendered template.